requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
passlib
pyjwt
ollama
faiss-cpu
python-dotenv
uvicorn
argon2-cffi
PyPDF2
python-docx
pandas
python-multipart
textual
rich
requests
pyinstaller
rank_bm25
python-telegram-bot
numpy
